bench/bench_in_faas/gcp_run/grpc/build_assets/requirements.txt
requests
boto3
tensorflow-serving-api==2.11.1
tensorflow==2.11.1
protobuf==3.19.6